requirements.txt
astroid==2.6.2
attrs==21.2.0
beautifulsoup4==4.9.3
certifi==2021.5.30
charset-normalizer==2.0.3
idna==3.2
iniconfig==1.1.1
isort==5.9.2
lazy-object-proxy==1.6.0
mccabe==0.6.1
packaging==21.0
pluggy==0.13.1
py==1.10.0
pylint==2.9.3
pymgclient==1.0.0
pyparsing==2.4.7
pytest==6.2.4
pytest-socket==0.4.0
requests==2.26.0
requests-mock==1.9.3
six==1.16.0
soupsieve==2.2.1
toml==0.10.2
urllib3==1.26.6
wrapt==1.12.1
